estate_crud/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
django-restframework==0.0.1
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
